Add integration tests
Turns out that the previous fix to optimise the build introduced a bug.... the unit tests need to run with full debug info, in order to get coverage info. So by saying "--no-build" we didn't have any built debug version of the assemblies we were testing. Need to have a rethink about all this.