Decouple core commands from the messages received by the api
These are different things - eg, core command for creating an environment includes the application id, but the message received by the api will get this from the url. Also think of stuff like user ids which will be from a token in the api, but included on the core command.