Fix failing tests
There was a deadlock when committing because the in-process handlers tried to read from the repository whilst the repository was still committing the events. Fixed this by making the in-process handlers just write to a stream, which is now processed by background services. This is more like how it will work in production, so all is good.