Fix dates for those DTOs that already have them
Problem was because of the way CQRSlite adds the timestamp to the event after the aggregate has been built, which resulted in the aggregate having the default timestamp until the system was restarted. Only solution I could think of was to not use the IEvent.Timestamp field to set these dates, and instead add our own OccurredAt property to the events which we set ourselves.